Revert "Trying Maps"
This reverts commit 9aaa44cb9b855b44a83d03828bd5669ab8961474.

+++ Header files/Bow.h
@@ -5,7 +5,7 @@ class Bow :
     public Item
 {
     float Damage;
-    float FireSpeed;
+    int FireSpeed;
     float FiringTimer;
     float timer;
     int frame;
